requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
requests==2.31.0
aiohttp==3.9.1
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
playwright==1.41.0
PyGithub==2.1.1
gitpython==3.1.41
jinja2==3.1.3
markdown==3.5.2
pandas==2.2.0
uuid7==0.1.0
cryptography==42.0.0
httpx==0.26.0
tenacity==8.2.3
openai==1.10.0
anthropic==0.18.1
beautifulsoup4==4.12.3
lxml==5.1.0
gradio==4.16.0
